Play system sound on volume changes (Fixes #4)
This commit introduces optional support for system sounds as audio
feedback when changing the volume with e.g. multimedia keys. This is
provided via libcanberra, either enabled or disabled at compile time.
Furthermore there is a setting in the properties dialog (FALSE by
default) to allow users to enable/disable this feature.

 Author:    Scott Shumate <scott@shumatech.com>