Use Source Output Info for recording indicator
This enables us to filter by application and ignore mixers like
pavucontrol, who monitor the recording channel to be able to show the
levels of the microphone.
Furthermore, by using the state tracked in the recording boolean we can
reduce the amount of redraws in the volume button (e.g. when changing
the volume with the mousewheel on hover) and thereby also reduce
flickering (see #81).