Do not override fatal log level
Maybe there was a reason for this override (the plugin generating too
many warnings and the fatal log level set too low by the panel in debug
mode), but this is not the case anymore. On the other hand, it can
override a lower log level set as fatal via G_DEBUG env var, and even
affect the panel if the plugin is loaded as internal (which is currently
the case on Wayland).

Related: xfce4-indicator-plugin!6
MR: !16