Do not set G_LOG_LEVEL_ERROR as fatal
There is no point in doing this since G_LOG_LEVEL_ERROR is always fatal.
On the other hand, it can override a lower log level set as fatal via
G_DEBUG env var, and even affect the panel if the plugin is loaded as
internal (which is currently the case on Wayland).