wnck: Use Libxfce4windowing when available
This makes the alternative raise method available on Wayland as well,
simplifying it a bit in the process, which should introduce no
difference in behavior on X11. The "wnck" prefix/suffix is generally
retained in variable and function names, as it is in the Xfconf property
name, and it doesn't seem necessary to rename everything at the cost of
having to deal with backwards compatibility.