Don't force set the default device
On some configurations we can get different default device for a while.
If we force-set the default device in this moment, we'll stuck with
different default device then we should have. For example on some
laptops connecting external microphone sets monitor as a default
recording device. In this case user have to switch back to microphone.

Closes: #66